+++ CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 <!-- insertion marker -->
+## Unreleased
+
+<small>[Compare with latest](https://github.com/ydethe/quizzy/compare/v0.2.2...HEAD)</small>
+
+### Added
+
+- Added healthcheck ([519e8a4](https://github.com/ydethe/quizzy/commit/519e8a45707dec0888ceefac49f9f8e85681131b) by Yann de The).
+
+<!-- insertion marker -->
 ## [v0.2.2](https://github.com/ydethe/quizzy/releases/tag/v0.2.2) - 2025-10-06
 
 <small>[Compare with v0.2.1](https://github.com/ydethe/quizzy/compare/v0.2.1...v0.2.2)</small>