+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- Added authentication on admin page ([a81dff1](https://github.com/ydethe/quizzy/commit/a81dff16f79485e9146ccfd6db5bdf33527353cd) by Yann de The).
 - Added Claim model ([069f01f](https://github.com/ydethe/quizzy/commit/069f01f10dd2bd77a0c265859c9f8b17b59b9661) by Yann de The).
 
 <!-- insertion marker -->