+++ CHANGELOG.md
@@ -6,9 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 <!-- insertion marker -->
-## Unreleased
+## [v0.2.4](https://github.com/ydethe/quizzy/releases/tag/v0.2.4) - 2025-10-10
 
-<small>[Compare with latest](https://github.com/ydethe/quizzy/compare/v0.2.3...HEAD)</small>
+<small>[Compare with v0.2.3](https://github.com/ydethe/quizzy/compare/v0.2.3...v0.2.4)</small>
 
 ### Added
 
@@ -19,7 +19,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Fixed dependencies ([28cc0a3](https://github.com/ydethe/quizzy/commit/28cc0a3bebae85ce7ea9351e868616eb35e26124) by Yann de The).
 
-<!-- insertion marker -->
 ## [v0.2.3](https://github.com/ydethe/quizzy/releases/tag/v0.2.3) - 2025-10-06
 
 <small>[Compare with v0.2.2](https://github.com/ydethe/quizzy/compare/v0.2.2...v0.2.3)</small>